[ty] split out type and annotation expressions
